Crate tzdata

tzdata uses the IANA Time Zone Database to provide conversion between timezones.

Reexports

pub use time::*;

Structs

Datetime
A precise point in time along associated to a Timezone.

Timezone
A named timezone from the IANA Time Zone Database