Struct tzdata::Datetime

pub struct Datetime<'a> {
    // some fields omitted
}

A precise point in time along associated to a Timezone.

The Datetime cannot be created on its own as it depends on a Timezone. tzdata does not expose a naive flavor of Datetime. To build a Datetime, instanciate a Timezone first and call the desired method.

let paris = tzdata::Timezone::new("Europe/Paris").unwrap();
let midnight_in_paris = paris.datetime(2015, 12, 25, 0, 0, 0, 0);

let utc = tzdata::Timezone::utc();
let t = midnight_in_paris.project(&utc);

assert_eq!(t.date(), (2015, 12, 24));
assert_eq!(t.time(), (23, 0, 0, 0));

Methods

impl<'a> Datetime<'a>

fn project<'b>(&self, tz: &'b Timezone) -> Datetime<'b>

Project the current Datetime in another Timezone.

fn date(&self) -> (i32, i32, i32)

Return the date component of the Datetime expressed in the associated Timezone. The tuple holds the year, month and day in this order.

fn time(&self) -> (i32, i32, i32, i32)

Return the time component of the Datetime expressed in the associated Timezone. The tuple holds the hour, minute, second and nanosecond in this order.