Struct tzdata::Deltatime

pub struct Deltatime(_);

A delta of time used in Datetime arithmetic.

This represents a duration in time and can be used to shift Datetime by a specified amount of time or as a result of comparing two Datetime. Different types of Deltatime can be created and each of those types affects Datetime differently. For example, Deltatime::seconds(86400) is different from Deltatime::days(1) because a Deltatime is dependant of the Datetime it applies to.

Example

let utc = tzdata::Timezone::utc();
let t = utc.datetime(2015, 6, 30, 0, 0, 0, 0);

let add_86400_secs = t + tzdata::Deltatime::seconds(86400);

assert_eq!(add_86400_secs.date(), (2015, 6, 30));
assert_eq!(add_86400_secs.time(), (23, 59, 60, 0));

Methods

impl Deltatime

fn nanoseconds(n: i64) -> Self

Create a delta of n nanoseconds. Possible leap seconds are accounted for.

fn microseconds(n: i64) -> Self

Create a delta of n microseconds. Possible leap seconds are accounted for.

fn milliseconds(n: i64) -> Self

Create a delta of n milliseconds. Possible leap seconds are accounted for.

fn seconds(n: i64) -> Self

Create a delta of n seconds. Possible leap seconds are accounted for.

fn minutes(n: i64) -> Self

Create a delta of n minutes. Possible leap seconds are accounted for.

fn hours(n: i64) -> Self

Create a delta of n hours. Possible leap seconds are accounted for.

fn days(n: i64) -> Self

Create a delta of n days. The time part of the shifted Datetime is not affected. This is a logical day, therefore, leap seconds are ignored.